Move non-x86 (host) specific parts of codegen-x86.c to codegen.c.
