+++ src/base/emu-i386/simx86/Makefile
@@ -11,7 +11,7 @@ include $(top_builddir)/Makefile.conf
 EM86DIR=$(REALTOPDIR)/src/emu-i386/simx86
 EM86FLG=-Dlinux -DDOSEMU
 ifeq ($(X86_JIT),1)
-JITFILES = codegen-x86.c fp87-x86.c sigsegv.c cpatch.c trees.c memory.c
+JITFILES = codegen.c codegen-x86.c fp87-x86.c sigsegv.c cpatch.c trees.c memory.c
 endif
 CFILES = interp.c cpu-emu.c modrm-gen.c $(JITFILES) \
 	codegen-sim.c fp87-sim.c modrm-sim.c protmode.c tables.c
