+++ src/base/emu-i386/simx86/codegen-x86.c
@@ -68,16 +68,6 @@
  * Clearly we hope that the 1st level cache of the host CPU works as
  * advertised ;-)
  *
- * There are two main functions here:
- *	AddrGen, which implements the AGU (Address Generation Unit).
- *		It calculates the address coming from ModRM and stores
- *		it into a well-defined register (edi in the x86 case)
- *	Gen, which does the ALU work and all the rest. There is no
- *		branch specific unit, as the branches are (in principle)
- *		all interpreted.
- * Both functions use a variable parameter approach, just to make them
- *	hard to follow ;-)
- *
  */
 
 /***************************************************************************
@@ -112,20 +102,9 @@
 #include <string.h>
 #include "utilities.h"
 #include "emu86.h"
-#include "misc/dlmalloc.h"
-#include "mapping/mapping.h"
 #include "codegen-x86.h"
-#include "cpatch.h"
-
-static void Gen_x86(int op, int mode, ...);
-static void AddrGen_x86(int op, int mode, ...);
-
-hitimer_u TimeStartExec;
-static TNode *LastXNode = NULL;
 
 /////////////////////////////////////////////////////////////////////////////
-
-#define	Offs_From_Arg()		(signed char)(va_arg(ap,int))
 
 /* This code is appended at the end of every instruction sequence. It
  * passes back the IP of the next instruction after the sequence.
@@ -160,30 +139,6 @@ static void _test_(void)
 
 /////////////////////////////////////////////////////////////////////////////
 
-static int goodmemref(dosaddr_t m)
-{
-	if (m < 0x110000) return 1;
-	if (dpmi_is_valid_range(m, 16))
-		return 1;
-	return 0;
-}
-
-/////////////////////////////////////////////////////////////////////////////
-
-
-void InitGen_x86(void)
-{
-	Fetch = jit_fetch_byte;
-	FetchW = jit_fetch_word;
-	FetchL = jit_fetch_dword;
-
-	Gen = Gen_x86;
-	AddrGen = AddrGen_x86;
-	UseLinker = USE_LINKER;
-}
-
-
-/////////////////////////////////////////////////////////////////////////////
 
 /* NOTE: parameters IG->px must be the last argument in a Gn() macro
  * because of the OR operator, which would cause trouble if the parameter
@@ -2448,441 +2403,6 @@ shrot0:
 }
 
 
-/*
- * address generator unit
- * careful - do not use eax, and NEVER change any flag!
- */
-static void AddrGen_x86(int op, int mode, ...)
-{
-	va_list	ap;
-	IMeta *I;
-	IGen *IG;
-#if PROFILE >= 2
-	hitimer_t t0 = 0;
-	if (debug_level('e')) t0 = GETTSC();
-#endif
-
-	if (CurrIMeta<0) {
-		CurrIMeta=0; InstrMeta[0].ngen=0; InstrMeta[0].flags=0;
-	}
-	I = &InstrMeta[CurrIMeta];
-	if (I->ngen >= NUMGENS) { leavedos_main(0xbac1); return; }
-	IG = &(I->gen[I->ngen]);
-	if (debug_level('e')>6) dbug_printf("AGEN: %3d %6x\n",op,mode);
-
-	va_start(ap, mode);
-	IG->op = op;
-	IG->mode = mode;
-
-	switch(op) {
-	case A_DI_0:			// base(32), imm
-	case A_DI_1: {			// base(32), {imm}, reg, {shift}
-		signed char ofs = (char)va_arg(ap,int);
-		signed char o;
-		IG->p0 = ofs;
-		IG->p1 = va_arg(ap,int);
-		if (op==A_DI_0)	break;
-		o = Offs_From_Arg();
-		IG->p2 = o;
-		}
-		break;
-	case A_DI_2: {			// base(32), {imm}, reg, reg, {shift}
-		signed char o1,o2;
-		signed char ofs = (char)va_arg(ap,int);
-		unsigned char sh;
-		IG->p0 = ofs;
-		IG->p1 = va_arg(ap,int);
-		o1 = Offs_From_Arg();
-		o2 = Offs_From_Arg();
-		IG->p2 = o1;
-		IG->p3 = o2;
-		sh = (unsigned char)(va_arg(ap,int));
-		IG->p4 = sh;
-		}
-		break;
-	case A_DI_2D: {			// modrm_sibd, 32-bit mode
-		signed char o;
-		signed char ofs = (char)va_arg(ap,int);
-		unsigned char sh;
-		IG->p0 = ofs;
-		IG->p1 = va_arg(ap,int);
-		o = Offs_From_Arg();
-		IG->p2 = o;
-		sh = (unsigned char)(va_arg(ap,int));
-		IG->p3 = sh;
-		}
-		break;
-	case A_SR_SH4: {	// real mode make base addr from seg
-		signed char o1 = Offs_From_Arg();
-		signed char o2 = Offs_From_Arg();
-		IG->p0 = o1;
-		IG->p1 = o2;
-		if (o1 == Ofs_SS)
-			I->flags |= F_INHI;
-		}
-		break;
-	}
-	va_end(ap);
-	I->ngen++;
-#if PROFILE >= 2
-	if (debug_level('e')) GenTime += (GETTSC() - t0);
-#endif
-}
-
-
-static void Gen_x86(int op, int mode, ...)
-{
-	int rcod=0;
-	va_list	ap;
-	IMeta *I;
-	IGen *IG;
-#if PROFILE >= 2
-	hitimer_t t0 = 0;
-	if (debug_level('e')) t0 = GETTSC();
-#endif
-
-	if (CurrIMeta<0) {
-		CurrIMeta=0; InstrMeta[0].ngen=0; InstrMeta[0].flags=0;
-	}
-	I = &InstrMeta[CurrIMeta];
-	if (I->ngen >= NUMGENS) leavedos_main(0xbac2);
-	IG = &(I->gen[I->ngen]);
-	if (debug_level('e')>6) dbug_printf("CGEN: %3d %6x\n",op,mode);
-
-	va_start(ap, mode);
-	IG->op = op;
-	IG->mode = mode;
-
-	switch(op) {
-	case L_NOP:
-	case L_CR0:
-	case L_ZXAX:
-	case L_DI_R1:
-	case S_DI:
-	case O_NOT:
-	case O_NEG:
-	case O_INC:
-	case O_DEC:
-	case O_MUL:
-	case O_CBWD:
-	case O_PUSH:
-	case O_PUSH1:
-	case O_PUSH2F:
-	case O_PUSH3:
-	case O_POP1:
-	case O_POP3:
-	case O_LEAVE:
-	case O_MOVS_MovD:
-	case O_MOVS_LodD:
-	case O_MOVS_StoD:
-	case O_MOVS_ScaD:
-	case O_MOVS_CmpD:
-	case O_RDTSC:
-	case O_INPDX:
-	case O_OUTPDX:
-		break;
-
-	case L_REG:
-	case S_REG:
-	case S_DI_R:
-	case L_LXS1:
-	case O_XCHG:
-	case O_CLEAR:
-	case O_TEST:
-	case O_SBSELF:
-	case O_BSWAP:
-	case O_XLAT:
-	case O_MOVS_SetA:
-	case O_MOVS_SavA:
-	case O_CMPXCHG: {
-		signed char o = Offs_From_Arg();
-		IG->p0 = o;
-		}
-		break;
-
-	case L_REG2REG:
-	case L_LXS2:	/* real mode segment base from segment value */
-	case O_XCHG_R: {
-		signed char o1 = Offs_From_Arg();
-		signed char o2 = Offs_From_Arg();
-		IG->p0 = o1;
-		IG->p1 = o2;
-		}
-		break;
-
-	case S_DI_IMM:
-	case L_IMM_R1:
-	case O_ADD_R:				// acc = acc op	reg
-	case O_OR_R:
-	case O_ADC_R:
-	case O_SBB_R:
-	case O_AND_R:
-	case O_SUB_R:
-	case O_XOR_R:
-	case O_CMP_R:
-	case O_INC_R:
-	case O_DEC_R:
-	case O_DIV:
-	case O_IDIV:
-	case O_PUSHI:
-	case JMP_TAILCODE: {
-		int v = va_arg(ap,int);
-		IG->p0 = v;
-		}
-		break;
-
-	case O_ADD_FR:				// reg = reg op	acc/imm
-	case O_OR_FR:
-	case O_ADC_FR:
-	case O_SBB_FR:
-	case O_AND_FR:
-	case O_SUB_FR:
-	case O_XOR_FR:
-	case O_CMP_FR: {
-		signed char o = Offs_From_Arg();
-		IG->p0 = o;
-		if (mode & IMMED) {
-			int v = va_arg(ap,int);
-			IG->p1 = v;
-		} }
-		break;
-
-	case L_IMM: {
-		signed char o = Offs_From_Arg();
-		int v = va_arg(ap,int);
-		IG->p0 = o;
-		IG->p1 = v;
-		}
-		break;
-
-	case L_MOVZS: {
-		signed char o;
-		rcod = (va_arg(ap,int)&1)<<3;	// 0=z 8=s
-		o = Offs_From_Arg();
-		IG->p0 = rcod;
-		IG->p1 = o;
-		}
-		break;
-
-	case O_IMUL:
-		if (mode&IMMED) {
-			int v = va_arg(ap,int);
-			signed char o = Offs_From_Arg();
-			IG->p0 = v;
-			IG->p1 = o;
-		}
-		if (!(mode&MBYTE)) {
-			if (mode&MEMADR) {
-				signed char o = Offs_From_Arg();
-				IG->p0 = o;
-			}
-		}
-		break;
-
-	case O_ROL:
-	case O_ROR:
-	case O_RCL:
-	case O_RCR:
-	case O_SHL:
-	case O_SHR:
-	case O_SAR:
-		if (mode & IMMED) {
-			unsigned char sh = (unsigned char)va_arg(ap,int);
-			IG->p0 = sh;
-		}
-		break;
-
-	case O_OPAX: {	/* used by DAA..AAD */
-			int n =	va_arg(ap,int);
-			IG->p0 = n;
-			// get n>0,n<3 bytes from parameter stack
-			IG->p1 = va_arg(ap,int);
-			if (n==2) IG->p2 = va_arg(ap,int);
-		}
-		break;
-
-	case O_POP:
-		if (mode & MRETISP) IG->p0 = va_arg(ap,int);
-		break;
-
-	case O_PUSH2:
-	case O_POP2: {
-		signed char o = Offs_From_Arg();
-		IG->p0 = o;
-		}
-		break;
-
-	case O_SLAHF:
-		rcod = va_arg(ap,int)&1;	// 0=LAHF 1=SAHF
-		IG->p0 = rcod;
-		break;
-
-	case O_SETFL:
-	case O_SETCC: {
-		unsigned char n = (unsigned char)va_arg(ap,int);
-		IG->p0 = n;
-		}
-		break;
-
-	case O_FOP:
-		I->flags |= F_FPOP;
-		// fall through
-	case O_INT: {
-		unsigned char exop = (unsigned char)va_arg(ap,int);
-		IG->p0 = exop;
-		IG->p1 = va_arg(ap,int);	// reg
-		}
-		break;
-
-	case O_BITOP: {
-		unsigned char n = (unsigned char)va_arg(ap,int);
-		signed char o = Offs_From_Arg();
-		IG->p0 = n;
-		IG->p1 = o;
-		} break;
-
-	case O_SHFD: {
-		unsigned char l_r = (unsigned char)va_arg(ap,int)&8;
-		signed char o = Offs_From_Arg();
-		IG->p0 = l_r;
-		IG->p1 = o;
-		if (mode & IMMED) {
-			unsigned char shc = (unsigned char)va_arg(ap,int)&0x1f;
-			IG->p2 = shc;
-		}
-		} break;
-
-	case JMP_INDIRECT:
-		break;
-
-	case JMP_LINK:		// opc, dspt, retaddr, link
-	case JLOOP_LINK: {
-		unsigned char opc = (unsigned char)va_arg(ap,int);
-		IG->p0 = opc;
-		IG->p1 = va_arg(ap,int);	// dspt
-		IG->p2 = va_arg(ap,int);	// dspnt
-		}
-		break;
-
-	case JF_LINK:
-	case JB_LINK: {		// opc, PC, dspt, dspnt, link
-		unsigned char opc = (unsigned char)va_arg(ap,int);
-		IG->p0 = opc;
-		IG->p1 = va_arg(ap,int);	// jpc
-		IG->p2 = va_arg(ap,int);	// dspt
-		IG->p3 = va_arg(ap,int);	// dspnt
-		}
-		break;
-
-	}
-
-	va_end(ap);
-	I->ngen++;
-#if PROFILE >= 2
-	if (debug_level('e')) GenTime += (GETTSC() - t0);
-#endif
-}
-
-
-/////////////////////////////////////////////////////////////////////////////
-
-
-static CodeBuf *ProduceCode(unsigned int PC, IMeta *I0)
-{
-	int i,j,nap,mall_req;
-	unsigned int adr_lo=0, adr_hi=0;
-	unsigned char *cp, *cp1, *BaseGenBuf, *CodePtr;
-	size_t GenBufSize;
-	CodeBuf *GenCodeBuf;
-
-	if (debug_level('e')>1) {
-	    e_printf("---------------------------------------------\n");
-	    e_printf("ProduceCode: CurrIMeta=%d\n",CurrIMeta);
-	}
-	if (CurrIMeta < 0) leavedos_main(0xbac3);
-
-	/* reserve space for auto-ptr and info structures */
-	nap = I0->ncount+1;
-
-	/* allocate the actual code buffer here; size is a worst-case
-	 * estimate based on measured bytes per opcode.
-	 *
-	 * Code buffer layout:
-	 *	0000	(GenCodeBuf) pointed from {TNode}.mblock
-	 *		contains a back pointer to the TNode
-	 * 0008/0004	self-pointer (address of this location)
-	 * 0010/0008	Addr2Pc table (nap) pointed from {TNode}.pmeta
-	 *	nap+10/8 actual code produced (BaseGenBuf)
-	 *		plus tail code
-	 * Only the code part is filled here.
-	 * GenBufSize contain a first guess of the amount of space required
-	 *
-	 */
-	GenBufSize = 0;
-	for (i=0; i<CurrIMeta; i++)
-	    GenBufSize += I0[i].ngen * MAX_GEND_BYTES_PER_OP;
-	mall_req = GenBufSize + offsetof(CodeBuf, meta) + sizeof(Addr2Pc) * nap + 32;// 32 for tail
-	GenCodeBuf = dlmalloc(mall_req);
-	/* actual code buffer starts from here */
-	BaseGenBuf = CodePtr = (unsigned char *)&GenCodeBuf->meta[nap];
-	I0->daddr = 0;
-	if (debug_level('e')>1)
-	    e_printf("CodeBuf=%p siz %zd CodePtr=%p\n",GenCodeBuf,GenBufSize,CodePtr);
-
-	for (i=0; i<CurrIMeta; i++) {
-	    IMeta *I = &I0[i];
-	    if (i==0) {
-		adr_lo = adr_hi = I->npc;
-	    }
-	    else {
-		if (I->npc < adr_lo) adr_lo = I->npc;
-		    else if (I->npc > adr_hi) adr_hi = I->npc;
-	    }
-	    cp = cp1 = CodePtr;
-	    I->daddr = cp - BaseGenBuf;
-	    for (j=0; j<I->ngen; j++) {
-		IGen *IG = &I->gen[j];
-		CodePtr = CodeGen(CodePtr, BaseGenBuf, IG);
-		if (CodePtr-cp1 > MAX_GEND_BYTES_PER_OP) {
-		    dosemu_error("Generated code (%zd bytes) overflowed into buffer, please "
-				 "increase MAX_GEND_BYTES_PER_OP=%d\n",
-				 CodePtr-cp1, MAX_GEND_BYTES_PER_OP);
-		    leavedos_main(0x535347);
-		}
-		if (debug_level('e')>1) {
-		    int dg = CodePtr-cp1;
-		    e_printf("PGEN(%02d,%02d) %3d %6x %2d %08x %08x %08x %08x %08x\n",
-			i,j,IG->op,IG->mode,dg,
-			IG->p0,IG->p1,IG->p2,IG->p3,IG->p4);
-		}
-		cp1 = CodePtr;
-	    }
-	    I->len = CodePtr - cp;
-	    if (debug_level('e')>3) GCPrint(cp, BaseGenBuf, I->len);
-	}
-	if (debug_level('e')>1)
-	    e_printf("Size=%td guess=%zd\n",(CodePtr-BaseGenBuf),GenBufSize);
-/**/ if ((CodePtr-BaseGenBuf) > GenBufSize) leavedos_main(0x535347);
-	if (PC < adr_lo) adr_lo = PC;
-	    else if (PC > adr_hi) adr_hi = PC;
-	I0->seqbase = adr_lo;
-	I0->seqlen  = adr_hi - adr_lo;
-
-	if (debug_level('e')>1)
-	    e_printf("---------------------------------------------\n");
-
-	I0->totlen = CodePtr - BaseGenBuf;
-
-	/* shrink buffer to what is actually needed */
-	mall_req = I0->totlen + offsetof(CodeBuf, meta) + sizeof(Addr2Pc) * nap;
-	GenCodeBuf = dlrealloc(GenCodeBuf, mall_req);
-	if (debug_level('e')>3)
-		e_printf("Seq len %#x:%#x\n",I0->seqlen,I0->totlen);
-
-	return GenCodeBuf;
-}
-
-
 /////////////////////////////////////////////////////////////////////////////
 /*
  * The node linker.
@@ -3022,7 +2542,7 @@ static void linknode(TNode *LG, TNode *G, linkdesc *L, unsigned target_type)
 	}
 }
 
-static void NodeLinker(TNode *LG, TNode *G)
+void NodeLinker(TNode *LG, TNode *G)
 {
 #if PROFILE >= 2
 	hitimer_t t0 = 0;
@@ -3156,105 +2676,6 @@ void NodeUnlinker(TNode *G)
 }
 
 
-/////////////////////////////////////////////////////////////////////////////
-/*
- * These are the functions which actually executes the generated code.
- *
- * There are two paths:
- * 1) for CloseAndExec_x86 we are ending a code generation phase, and our code
- *	is still in the CodeBuf together with all its detailed info stored
- *	in InstrMeta. First we close the sequence adding the TailCode;
- *	it, and move it to the collecting tree and clear the temporary
- *	structures. Then, in Exec_x86 we execute the code.
- *	The PC parameter is the address in the source code of the next
- *	instruction following the end of the code block. It will be stored
- *	into the TailCode of the block.
- * 2) We are executing a sequence found in the collecting tree.
- *	Exec_x86 is called directly.
- *	G is the node we found (possibly the start of a chain of linked
- *	code sequences).
- *
- * When the code is executed, it returns in eax the source address of the
- * next instruction to find/parse.
- *
- */
-
-TNode *Close_x86(unsigned int PC, int mode)
-{
-	IMeta *I0;
-	TNode *G;
-	CodeBuf *GenCodeBuf;
-
-	if (CurrIMeta <= 0) {
-/**/		e_printf("(X) Nothing to exec at %08x\n",PC);
-		return NULL;
-	}
-
-	// we're creating a new node
-	I0 = &InstrMeta[0];
-
-	if (debug_level('e')>2) {
-		e_printf("==== Closing sequence at %08x\n", PC);
-	}
-
-	GenCodeBuf = ProduceCode(PC, I0);
-	/* check for fatal error */
-	if (TheCPU.err < 0)
-		return NULL;
-
-	NodesParsed++;
-#if PROFILE
-	if (debug_level('e')) TotalNodesParsed++;
-#endif
-	G = Move2Tree(I0, GenCodeBuf);		/* when is G==NULL? */
-	/* InstrMeta will be zeroed at this point */
-	/* mprotect the page here; a page fault will be triggered
-	 * if some other code tries to write over the page including
-	 * this node */
-	e_markpage(G->seqbase, G->seqlen);
-	G->cs = LONG_CS;
-	G->mode = mode;
-	/* check links INSIDE current node */
-	if (0 == (EFLAGS & EFLAGS_TF) ) {
-		NodeLinker(G, G);
-	}
-	return G;
-}
-
-#ifdef __i386__
-  __attribute__((target("sse")))
-#endif
-static inline void prefetch(unsigned char *p)
-{
-	__builtin_prefetch(p);
-}
-
-static unsigned int Exec_x86_pre(unsigned char *ecpu)
-{
-	unsigned long flg;
-
-	/* get the protected mode flags. Note that RF and VM are cleared
-	 * by pushfd (but not by ints and traps) */
-	flg = getflags();
-
-	/* pass TF=0, IF=1, DF=0 */
-	flg = (flg & ~(EFLAGS_CC|EFLAGS_IF|EFLAGS_DF|EFLAGS_TF)) |
-	       (EFLAGS & EFLAGS_CC) | EFLAGS_IF;
-
-#ifdef __i386__
-	if (config.cpuprefetcht0)
-#endif
-		prefetch(ecpu);
-
-	return flg;
-}
-
-static void Exec_x86_post(unsigned long flg, unsigned int mem_ref)
-{
-	EFLAGS = (EFLAGS & ~EFLAGS_CC) | (flg &	EFLAGS_CC);
-	TheCPU.mem_ref = mem_ref;
-}
-
 /* stack frame for compiled code:
  * esp+00	TheCPU flags
  *     04/08	return address
@@ -3271,7 +2692,7 @@ static void Exec_x86_post(unsigned long flg, unsigned int mem_ref)
 #define R_REG(r) "%e"#r
 #define EXEC_CLOBBERS
 #endif
-static unsigned Exec_x86_asm(unsigned *mem_ref, unsigned long *flg,
+unsigned Exec_x86_asm(unsigned *mem_ref, unsigned long *flg,
 		unsigned char *ecpu, unsigned char *SeqStart)
 {
 	unsigned ePC;
@@ -3331,7 +2752,7 @@ static unsigned Exec_x86_asm(unsigned *mem_ref, unsigned long *flg,
 	return ePC;
 }
 
-static unsigned Exec_x86_asm_fpu(unsigned *mem_ref, unsigned long *flg,
+unsigned Exec_x86_asm_fpu(unsigned *mem_ref, unsigned long *flg,
 		unsigned char *ecpu, unsigned char *SeqStart,
 		unsigned short seqflg)
 {
@@ -3363,181 +2784,6 @@ static unsigned Exec_x86_asm_fpu(unsigned *mem_ref, unsigned long *flg,
 		}
 	}
 	return ePC;
-}
-
-unsigned int Exec_x86(TNode *G)
-{
-	unsigned long flg;
-	unsigned char *ecpu;
-	unsigned int mem_ref;
-	unsigned int ePC;
-	unsigned short seqflg = G->flags;
-	unsigned char *SeqStart = G->addr;
-
-	ecpu = CPUOFFS(0);
-	if (debug_level('e')>1) {
-		if (sigalrm_pending()>0) e_printf("** SIGALRM is pending\n");
-		e_printf("==== Executing code at %p flg=%04x\n",
-			SeqStart,seqflg);
-	}
-#ifdef ASM_DUMP
-	fprintf(aLog,"%p: exec\n",G->key);
-#endif
-#if PROFILE >= 2
-	hitimer_t TimeStartExec;
-	if (debug_level('e')) TimeStartExec = GETTSC();
-#endif
-	flg = Exec_x86_pre(ecpu);
-	ePC = Exec_x86_asm_fpu(&mem_ref, &flg, ecpu, SeqStart, seqflg);
-	Exec_x86_post(flg, mem_ref);
-
-	if (debug_level('e')) {
-#if PROFILE >= 2
-	    ExecTime += GETTSC() - TimeStartExec;
-#endif
-	    if (debug_level('e')>1) {
-		e_printf("** End code, PC=%08x sig=%x\n",ePC,
-		    sigalrm_pending());
-		if ((debug_level('e')>3) && (seqflg & F_FPOP)) {
-		    e_printf("  %s\n", e_trace_fp());
-		}
-		/* DANGEROUS - can crash dosemu! */
-		if ((debug_level('e')>4) && goodmemref(mem_ref)) {
-		    e_printf("*mem_ref [%#08x] = %08x\n",mem_ref,
-			     READ_DWORD(mem_ref));
-		}
-	    }
-	}
-	/* signal_pending at this point is 1 if there was ANY signal,
-	 * not just a SIGALRM
-	 */
-	if ((G->flags & F_INHI) && !(G->seqnum == 1 && (CEmuStat & CeS_INHI))) {
-		/* ignore signals and traps for movss/popss; if there is just
-		   one compiled instruction it should be ignored unconditionally
-		   if signals and traps were already ignored */
-		CEmuStat |= CeS_INHI;
-		CEmuStat &= ~CeS_TRAP;
-	} else {
-		CEmuStat &= ~(CeS_INHI|CeS_MOVSS);
-		if (sigalrm_pending()) {
-			CEmuStat|=CeS_SIGPEND;
-			sigalrm_pending_w(0);
-		}
-	}
-
-#if defined(SINGLESTEP)
-	InvalidateNodeRange(G->key, 1, NULL);
-	avltr_delete(G->key);
-	if (debug_level('e')>1) e_printf("\n%s",e_print_regs());
-#else
-	/*
-	 * After execution comes the linker stage.
-	 * So the order is:
-	 *	1) build code sequence in the IMeta buffer
-	 *	2) move buffer to a newly allocated node in the tree
-	 *	3) execute it, always returning back at the end
-	 *	4) link it to other nodes
-	 * Linking: a node is linked with the next
-	 * one in execution order, provided that the end source address
-	 * of the preceding node matches the start source address of the
-	 * following (i.e. no interpreted instructions in between).
-	 */
-	if (G && G->alive>0) {
-		/* check links FROM LastXNode TO current node */
-		if (LastXNode && LastXNode->alive > 0)
-			NodeLinker(LastXNode, G);
-		if (debug_level('e')>2 && G != LastXNode)
-			e_printf("New LastXNode=%08x\n",G->key);
-		LastXNode = G;
-	}
-#endif
-
-	return ePC;
-}
-
-/* fast loop, only used if nothing special is going on; if anything
-   out of the ordinary happens, the above Exec_x86() is called */
-unsigned int Exec_x86_fast(TNode *G)
-{
-	unsigned char *ecpu = CPUOFFS(0);
-	unsigned long flg = Exec_x86_pre(ecpu);
-	unsigned int ePC, mem_ref;
-	unsigned mode = G->mode;
-
-	do {
-		ePC = Exec_x86_asm(&mem_ref, &flg, ecpu, G->addr);
-		if (G->alive > 0 && LastXNode && LastXNode->alive > 0) {
-			if (LastXNode->unlinked_jmp_targets &&
-			    (LastXNode->clink_t.target == G->key ||
-			     LastXNode->clink_nt.target == G->key))
-				NodeLinker(LastXNode, G);
-			LastXNode = G;
-		}
-		if (sigalrm_pending()) {
-			CEmuStat|=CeS_SIGPEND;
-			break;
-		}
-	} while (!TheCPU.err2 && (G=FindTree(ePC)) &&
-		 GoodNode(G, mode) && !(G->flags & (F_FPOP|F_INHI)));
-
-	Exec_x86_post(flg, mem_ref);
-	sigalrm_pending_w(0);
-	return ePC;
-}
-
-#define fB(c,x) (((uint8_t *)(c))[(x) & ~_PAGE_MASK])
-static inline uint16_t fW(void *c, dosaddr_t x)
-{
-	if (x & 1)
-		return ((fB(c, x + 1) << 8) | fB(c, x));
-	return (((uint16_t *)(c))[(x & ~_PAGE_MASK) >> 1]);
-}
-static inline uint32_t fD(void *c, dosaddr_t x)
-{
-	if (x & 3)
-		return (((uint32_t)fW(c, x + 2) << 16) | fW(c, x));
-	return (((uint32_t *)(c))[(x & ~_PAGE_MASK) >> 2]);
-}
-
-uint8_t jit_fetch_byte(dosaddr_t x)
-{
-	void *cache[2];
-
-	e_fetch(x, 1, cache);
-	return fB(cache[0], x);
-}
-
-uint16_t jit_fetch_word(dosaddr_t x)
-{
-	void *cache[2];
-
-	if ((x & ~CGRMASK) != ((x + 1) & ~CGRMASK))
-		return ((jit_fetch_byte(x + 1) << 8) | jit_fetch_byte(x));
-	e_fetch(x, 2, cache);
-	if ((x & _PAGE_MASK) != ((x + 1) & _PAGE_MASK))
-		return ((fB(cache[1], x + 1) << 8) | fB(cache[0], x));
-	return fW(cache[0], x);
-}
-
-uint32_t jit_fetch_dword(dosaddr_t x)
-{
-	void *cache[2];
-
-	if ((x & ~CGRMASK) != ((x + 3) & ~CGRMASK))
-		return (((uint32_t)jit_fetch_word(x + 2) << 16) |
-				jit_fetch_word(x));
-	e_fetch(x, 4, cache);
-	if ((x & _PAGE_MASK) != ((x + 3) & _PAGE_MASK)) {
-		if (!(x & 1))
-			return (((uint32_t)fW(cache[1], x + 2) << 16) |
-					fW(cache[0], x));
-		if (x & 2)
-			return (((fD(cache[1], x + 1) & 0xffffff) << 8) |
-					fB(cache[0], x));
-		return (((uint32_t)fB(cache[1], x + 3) << 24) |
-					(fD(cache[0], x - 1) >> 8));
-	}
-	return fD(cache[0], x);
 }
 
 /////////////////////////////////////////////////////////////////////////////
