+++ src/base/emu-i386/simx86/codegen-x86.h
@@ -125,8 +125,14 @@ unsigned int Exec_x86_fast(TNode *G);
 //
 unsigned char *Fp87_op_x86(unsigned char *CodePtr, int exop, int reg);
 void InitGen_x86(void);
+void NodeLinker(TNode *LG, TNode *G);
 void NodeUnlinker(TNode *G);
 unsigned char *CodeGen(unsigned char *CodePtr, unsigned char *BaseGenBuf, const IGen *IG);
+unsigned Exec_x86_asm(unsigned *mem_ref, unsigned long *flg,
+		      unsigned char *ecpu, unsigned char *SeqStart);
+unsigned Exec_x86_asm_fpu(unsigned *mem_ref, unsigned long *flg,
+			  unsigned char *ecpu, unsigned char *SeqStart,
+			  unsigned short seqflg);
 
 static __inline__ int GoodNode(TNode *G, int mode)
 {
